refactored file structure, removed dead code
