+++ src/components/Editor/SideBar/ApparatusListBox.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import Button from "@mui/material/Button";
 import styled from "styled-components";
-import { getAssetBundles } from "../../util/jsonParsing";
-import { SerializedApparatus } from "../../util/types";
+import { getAssetBundles } from "../../../util/jsonParsing";
+import { SerializedApparatus } from "../../../util/types";
 
 const Box = styled.div`
   background: #fffaf0;
