+++ src/components/Editor/SideBar/SideBarItem.tsx
@@ -5,11 +5,11 @@ import StyleIcon from "@mui/icons-material/Style";
 import FormatListBulletedIcon from "@mui/icons-material/FormatListBulleted";
 import SaveAltIcon from "@mui/icons-material/SaveAlt";
 import KeyboardReturnIcon from "@mui/icons-material/KeyboardReturn";
-import { SideBarContext } from "../../util/customHooks/SideBarContext";
-import { saveExp } from "../../util/cloudOperations/writeToCloud";
-import { useActionList } from "../../util/customHooks/overlayfunc";
-import { urlFor } from "../../util/utils";
-import { GlobalContext, globalContextTypes } from "../../../pages/experience";
+import { SideBarContext } from "../../../util/customHooks/SideBarContext";
+import { saveExp } from "../../../util/cloudOperations/writeToCloud";
+import { useActionList } from "../../../util/customHooks/overlayfunc";
+import { urlFor } from "../../../util/utils";
+import { GlobalContext, globalContextTypes } from "../../../../pages/experience";
 
 const SideBarItemBox = styled.div`
   position: flex;
