+++ src/components/Editor/WebglBox.tsx
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 import Unity from "react-unity-webgl";
 import { useContext, useEffect, useState } from "react";
-import Loading from "./Loading";
-import { unityContext, loadApparatus } from "../util/unityContextActions";
-import Overlay2 from "./UIComponent/Overlay2";
-import { GlobalContext, globalContextTypes } from "../../pages/experience";
+import Loading from "../Loading";
+import { unityContext, loadApparatus } from "../../util/unityContextActions";
+import Overlay from "./Overlay";
+import { GlobalContext, globalContextTypes } from "../../../pages/experience";
 
 const WebglRoot = styled.div`
   display: relative;
@@ -48,7 +48,7 @@ function WebglBox({ userId }: WebglProps): JSX.Element {
         }}
         tabIndex={1}
       />
-      <Overlay2 />
+      <Overlay />
     </WebglRoot>
   ) : (
     <WebglRoot>
