+++ src/test/actionSequenceBox.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import React from "react";
 import { getActions, getAssetBundles } from "../util/jsonParsing";
-import ActionSequenceBox from "../components/editorBoxes/ActionSequenceBox";
+import ActionSequenceBox from "../components/Editor/SideBar/ActionSequenceBox";
 import {testmetadata1} from "../util/testConstants";
 
 test("ActionSequenceBox renders without crashing", () => {
