+++ src/test/apparatusListBox.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import React from "react";
-import ApparatusListBox from "../components/editorBoxes/ApparatusListBox";
+import ApparatusListBox from "../components/Editor/SideBar/ApparatusListBox";
 import {testmetadata1} from "../util/testConstants";
 
 test("ApparatusListBox renders without crashing", () => {
